Move binary options to its own module
